distutils3.bbclass: add egg files/directories to python path
Packages that use .egg files or directories for installation may
not be found when imported at the python3 interpreter.
.egg files/directories path must be included in a .pth file to
be appropriately included in python path.

This change looks for .egg files/directories in sitepackages
and adds its path to a .pth file during package installation.
It ensures that any new package that uses .egg files/recipes
will be appropriately added to path by performing the check from
the distutils3 class.

[YOCTO #8673]

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>