git-series: add the new series-pull-request
Managing multiple versions/revisions of a patch with the git-series
command involves several manual steps, providing a poor user experience
due to the required learning and complexity to use it.

The new series-pull-request script allows users to manage multiple patch
revisions following almost the same procedure as the one used with the
currently used create-pull-request script.

This new script includes a cover letter and the series patches as usual,
plus the revision number in prefix, preserves each revision patches in a
unique branch, allows user to point out the revision changes in a new
cover letter, using the same syntax as the create-pull-request (with the
only additon of "-v" option to indicate a revision number), and provides
patches that can be sent  with the existing send-pull-request script.

[YOCTO Innovation Exploration]

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>