oeqa/runtime/context.py: Add automatic target loading
This adds automatic target loading for runtime testing. It will look for
targets in 'lib/oeqa/core/target' in every layer used in the build.

Introspection was used to achieve this functionality, the idea is to load all
target modules and they will be registered with registerTarget decorator, so
it is less error prone than previous implementation.

testimage and testexport bbclasses were modified to support this. testexport
required more work to find the targets and copy them into the exporting
directory. The executor also needed to remove current choices (simpleremote,
qemu) because now we can have custom targets.

Signed-off-by: Mariano Lopez <mariano.lopez@linux.intel.com>