useradd.bbclass: Make postinst functions verbose if BB_VERBOSE_LOGS is set
If BB_VERBOSE_LOGS is enabled, then make the postinst functions use
set -x so that their execution is traced. This greatly simplifies
debugging, especially in case of failures as nothing is recorded in
the logs otherwise to indicate what caused the failure.

Also reduce code duplication in useradd_sysroot_sstate() by only
having one copy of the code to create the postinst files.

Signed-off-by: Peter Kjellerstedt <peter.kjellerstedt@axis.com>