reproducible_build.bbclass: initial support for binary reproducibility
Conditionally set some environment variables in order to achieve
improved binary reproducibility. Providing BUILD_REPRODUCIBLE_BINARIES is
set to "1", we set the following environment variables:

export PYTHONHASHSEED=0
export PERL_HASH_SEED=0
export TZ="UTC"

We export and set SOURCE_DATE_EPOCH. The value for this variable is obtained
after source code for a recipe has been unpacked, but before it is patched.
If the code comes from a GIT repo, we get the timestamp from the top
commit. Otherwise, if present, we get mtime from known files suche as NEWS,
ChangeLog, etc. If this fails, we go through all files and get the timestamp
from the youngest one. We create an individual timestamp for each recipe.
The timestamp is stored in the file 'src_date_epoch.txt'. Later on, each task
reads this file and sets SOURCE_DATE_EPOCH based on the value found in the file.

[YOCTO#11178]
[YOCTO#11179]

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>