image_types_wic: schedule prepare_wic_build correctly
When EFI is enabled we need to ensure that kernel and bootloader
artifacts are fully deployed before running prepare_wic_build.
This is done by scheduling this task after do_deploy task of
EFI_PROVIDER class and by adding dependencty to virtual/kernel:do_deploy.

Scheduled prepare_wic_build only if wic build is enabled to
avoid running it when it's not needed.

Signed-off-by: Ed Bartosh <ed.bartosh@linux.intel.com>