tinfoil: add functionality for running full builds
Up to this point, if you wanted to run build tasks in the normal way
they get run, there was no other way than to shell out to bitbake.
As long as we're prepared to handle the events produced, we can create a
wrapper around calling the buildTargets command. Borrow code from knotty
to do this so that we get the expected running task display (courtesy of
TermFilter) and Ctrl+C handling.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>