recipetool: refactor code for ensuring npm is available
Across devtool and recipetool we had an ugly set of code for ensuring
that we can call an npm binary, and much of that ugliness was a result
of not being able to run build tasks when tinfoil was active - if
recipetool found that npm was required and we didn't know beforehand
(e.g. we're fetching from a plain git repository as opposed to an npm://
URL where it's obvious) then it had to exit and return a special result
code, so that devtool knew it needed to build nodejs-native and then
call recipetool again. Tinfoil now allows us to run build tasks
directly, so we can take advantage of that instead and throw away the
ugly code at the same time.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>